+++ CHANGELOG.md
@@ -123,3 +123,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Contact and Subscribe forms are saving Name on Mailchimp
 - Fixed Who We Are section on mobile.
 - Allow `BlogPostsContainer` display a dynamic number of cards per row with swipe functionality
+- Fixed search funcitionality on blog page